refactor: rename drawer package to settings
